requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Limiter
flask-smorest
marshmallow
marshmallow-sqlalchemy
python-dotenv
PyMySQL
